requirements.txt
asgiref==3.4.1
attrs==20.3.0
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
constantly==15.1.0
coverage==6.2
cryptography==3.3.1
cssselect==1.1.0
Django==3.2.12
django-admin-autocomplete-filter==0.7.1
django-filter==21.1
djangorestframework==3.12.2
happybase==1.2.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
install==1.3.5
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
logzero==1.6.3
lxml==4.6.2
parsel==1.6.0
ply==3.11
Protego==0.1.16
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pyquery==1.4.3
python-dotenv==0.19.2
pytz==2020.5
queuelib==1.5.0
redis==3.5.3
requests==2.25.1
Scrapy==2.4.1
scrapy-djangoitem==1.1.1
scrapy-redis==0.6.8
scrapyd==1.2.1
scrapyd-client==1.1.0
service-identity==18.1.0
six==1.15.0
SQLAlchemy==1.3.23
sqlparse==0.4.1
thriftpy2==0.4.14
tld==0.12.5
Twisted==20.3.0
typing_extensions==4.1.1
urllib3==1.26.3
w3lib==1.22.0
zope.interface==5.2.0
